# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "First Laravel API",
        "description": "Simple Laravel Api who manage users and associated profiles",
        "version": "0.0.1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user and generate a token",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#components/schemas/token"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log out and delete user token",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid user"
                    }
                }
            }
        },
        "/api/profile/index": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "List all profiles.",
                "operationId": "69bf9ef74089f011e004a0017c4ddf11",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit of results returned",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "orderby",
                        "in": "query",
                        "description": "column to order data",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "direction of data ordenation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/profile"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "0": "bearerAuth:{}"
                    }
                ]
            }
        },
        "/api/profile/store": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Create a new profile",
                "operationId": "f8d0fdc840e2f4f6572e656002f37dde",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#components/schemas/profile"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#components/schemas/profile"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "0": "bearerAuth:{}"
                    }
                ]
            }
        },
        "/api/profile/{id}/show": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Search and show a profile by id.",
                "operationId": "21ef95c6363421071b09c8ac717f6394",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#components/schemas/profile"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "0": "bearerAuth:{}"
                    }
                ]
            }
        },
        "/profile/{id}/update": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update an existing profile",
                "operationId": "d6fcbec23f0daf4cee0c50c2c3c4c1fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#components/schemas/profile"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "0": "bearerAuth:{}"
                    }
                ]
            }
        },
        "/api/profile/{id}/destroy": {
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Delete a profile by id.",
                "operationId": "44e1f66c1081db123b8a1c2957caadb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "0": "bearerAuth:{}"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "profile": {
                "title": "profile",
                "description": "profile schema",
                "required": [
                    "user_id",
                    "profile"
                ],
                "properties": {
                    "user_id": {
                        "description": "Id of the user who own the profile",
                        "type": "integer",
                        "format": "int32"
                    },
                    "profile": {
                        "description": "Type of profile",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token to authenticate users",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
}